Separate _multiufuncs out of _basic
