+++ scipy/special/meson.build
@@ -338,10 +338,11 @@ python_sources = [
   '_add_newdocs.py',
   '_basic.py',
   '_ellip_harm.py',
+  '_input_validation.py',
   '_lambertw.py',
   '_logsumexp.py',
   '_mptestutils.py',
-  '_multiufunc.py',
+  '_multiufuncs.py',
   '_orthogonal.py',
   '_orthogonal.pyi',
   '_sf_error.py',
